Version.004
Added working Joystick Module connected to an Arduino Nano as a proof of concept that the game will be able to be played.
Altered some parameters so that joystick will not break any features eg. drag.
Added and will be using the ArduinoInput.cs file to control inputs from the Joystick.