Version.006
Added Checkpoints throughout the track that when the car collides with in the correct order will be the basis for completing a lap. Will be able to finish race until all checkpoints are driven through in consecutive order.
Added CarLapCounter to include the code with the checkpoints, as well as count how many laps the player has done and reset the checkpoints after each lap has been completed. This also checks when the race is over, and informs the TopDownCarController.
The TopDownCarController now knows when to stop the vehicle when the race finishes.
PositionHandler has been started in case we want to impliment multiplayer, but for now it simply checks when we pass every checkpoint using OnPassCheckpoint(). This could be moved to CarLapCounter in the future.
Also included Arduino code to get the joystick to work in case I hadn't already.